+++ test/cmd/sample-metrics-plugin/internal/plugin/plugin.go
