+++ test/cmd/sample-metrics-plugin/internal/plugin/plugin_test.go
